compensate for paranoid peopl that have fopen turned off
git-svn-id: https://plugins.svn.wordpress.org/wp-security-scan/trunk@40658 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ securityscan.php
@@ -4,7 +4,7 @@ Plugin Name: WP Security Scan
 Plugin URI: http://semperfiwebdesign.com/plugins/wp-security-scan/
 Description: Perform security scan of WordPress installation.
 Author: Michael Torbert
-Version: 2.2.56.24
+Version: 2.2.56.25
 Author URI: http://semperfiwebdesign.com/
 */
 
@@ -122,12 +122,15 @@ if (file_exists($filename)) {
 <br /><br />
 <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=mrtorbert%40gmail%2ecom&item_name=Support%20WordPress%20Security%20Scan%20Plugin&no_shipping=0&no_note=1&tax=0&currency_code=USD&lc=US&bn=PP%2dDonationsBF&charset=UTF%2d8" target="_blank">Click here to support this plugin.</a>
 <br /><br /><h4>Recent Donations</h4></div><?php 
-$file = fopen("http://semperfiwebdesign.com/donations.php", "r");
-while ( ! feof($file) ) {
-      $text = fgets($file);
-      echo '<li>' . $text . '</li>';
+
+if (function_exists('fopen')) {
+   $file = fopen("http://semperfiwebdesign.com/donations.php", "r");
+   while ( ! feof($file) ) {
+         $text = fgets($file);
+         echo '<li>' . $text . '</li>';
+   }
+   fclose($file);
 }
-fclose($file);
 ?>
 </div>
 <div style="clear:both"></div>
